[feature] create JwtAuthenticationConverter bean to correct retrieve User role from Auth0 jwt
